Relocate fetch_coastlines() to assets directory.
