+++ src/geovista/cache.py
@@ -16,14 +16,11 @@ from importlib.resources import files
 import os
 
 import pooch
-import pyvista as pv
 
-from .common import COASTLINES_RESOLUTION
 from .config import resources
 
 __all__ = [
     "CACHE",
-    "fetch_coastlines",
     "pooch_mute",
     "reload_registry",
 ]
@@ -70,39 +67,6 @@ CACHE.load_registry(
 GEOVISTA_POOCH_MUTE: bool = (
     os.environ.get("GEOVISTA_POOCH_MUTE", "false").lower() == "true"
 )
-
-
-def fetch_coastlines(resolution: str | None = None) -> pv.PolyData:
-    """Get the Natural Earth coastlines for the required resolution.
-
-    If the resource is not already available in the geovista :data:`CACHE`,
-    then it will be downloaded from the :data:`BASE_URL`.
-
-    Parameters
-    ----------
-    resolution : str, optional
-        The resolution of the Natural Earth coastlines, which may be either
-        ``110m``, ``50m`` or ``10m``. Defaults to
-        :data:`geovista.common.COASTLINES_RESOLUTION`.
-
-    Returns
-    -------
-    PolyData
-        The coastlines mesh.
-
-    Notes
-    -----
-    .. versionadded:: 0.1.0
-
-    """
-    if resolution is None:
-        resolution = COASTLINES_RESOLUTION
-
-    fname = f"ne_coastlines_{resolution}.vtk"
-    processor = pooch.Decompress(method="auto", name=fname)
-    resource = CACHE.fetch(f"natural_earth/physical/{fname}.bz2", processor=processor)
-
-    return pv.read(resource)
 
 
 def pooch_mute(silent: bool = True) -> None:
