+++ src/geovista/geometry.py
@@ -20,7 +20,7 @@ import numpy as np
 import pyvista as pv
 from shapely import LineString, MultiLineString
 
-from .cache import fetch_coastlines
+from . import fetch_coastlines
 from .common import (
     COASTLINES_RESOLUTION,
     GV_FIELD_RADIUS,
